requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# PubMed
biopython==1.85

# Google Patents (BigQuery)
google-cloud-bigquery==3.27.0

# Web search & scraping
duckduckgo-search==8.1.1
beautifulsoup4==4.12.3
httpx==0.28.1

# LLM extraction
openai==1.58.1

# Excel export
pandas==2.2.3
openpyxl==3.1.5
